+++ src/GUI/Views/MainViewControl.xaml.cs
@@ -103,6 +103,7 @@ public partial class MainViewControl : MainViewControlViewBase
 			[nameof(AppKeys.ToggleFilterFocus)] = ("Redux.Icon.Funnel", true, null),
 			[nameof(AppKeys.DeleteSelectedMods)] = ("Redux.Icon.Trash", true, "ReduxErrorBrush"),
 			[nameof(AppKeys.ToggleModFileNames)] = ("Redux.Icon.DocumentText", true, null),
+			[nameof(AppKeys.ToggleOverrideMods)] = ("Redux.Icon.ChevronUpStroke", true, null),
 			[nameof(AppKeys.OpenPreferences)] = ("Redux.Icon.Settings", true, null),
 			[nameof(AppKeys.OpenThemeAppearance)] = ("Redux.Icon.ColorPalette", true, null),
 			[nameof(AppKeys.OpenKeybindings)] = ("Redux.Icon.Key", true, null),